Define mg.Dependency interface (#101)
Target functions are used with mg.Deps and mg.ContextDeps to list
dependencies.  This can result in lots of simple internal targets
like:

    func updateTestServer1() { updateTestServer(1) }

This commit introduces a new Dependency interface that mg.Deps and
mg.ContextDeps can detect and use instead of its internal map of
functions to sync.Once.  It also introduces a means of converting
existing targets to dependencies using a similar map.