Use Target Names Instead of Function Addresses (#101)
Using function addresses to map functions to sync.Once instances
lets us use anonymous functions as targets with mg.Deps.  This
seemed great to me until I thought about how we would explain
that the address of a closure is fixed at compile time, not run
time.

This commit changes makeDependency to use the name of the function
like the current behavior of mg.Deps and mg.ContextDeps.